Working version of lightdash
**Summary**
I have just got back to the project and I have just got lightdash working. I had previously used the default lightdash image but then changed in to a custom image from the docker file. This was unsuccessful and I have now changed it back to the default image.

**Additions/Removals**
Minor changes to the docker-compose.yaml file and minor other changes.

**Additional Notes**
Not applicable